[Refractor] move time formatter to formatters
